# apps/landing/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FortyOne Public API",
    "version": "1.0.0",
    "description": "Machine-readable contract for stable FortyOne discovery, health, feedback, and the user-authorized remote MCP endpoint.",
    "contact": {
      "name": "FortyOne",
      "email": "hello@complexus.tech",
      "url": "https://www.fortyone.app/contact"
    }
  },
  "servers": [
    { "url": "https://api.fortyone.app", "description": "Production" }
  ],
...
  "paths": {
...
    "/readiness": {
      "get": {
        "operationId": "getAPIReadiness",
        "summary": "Check API readiness",
        "description": "Confirms whether the API and its database dependency are ready to serve requests.",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "API is ready",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/mcp": {
      "post": {
        "operationId": "sendMCPMessage",
        "summary": "Send an MCP JSON-RPC message",
        "description": "Authenticated Streamable HTTP endpoint for permission-aware workspace, planning, creation, and delivery-analysis tools.",
        "tags": ["Discovery"],
        "security": [{ "oauth2": ["mcp:access"] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JSONRPCRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "401": { "description": "OAuth bearer token required" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.fortyone.app/oauth/authorize",
            "tokenUrl": "https://api.fortyone.app/oauth/token",
            "scopes": { "mcp:access": "Use FortyOne MCP tools as the connected user" }
          }
        }
      }
    },
...
    "responses": {
      "Error": {
        "description": "Structured error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": { "description": "Operation-specific response data" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["data", "error"],
        "properties": {
          "data": { "type": "null" },
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": ["code", "message", "hint"],
        "properties": {
          "code": { "type": "string", "examples": ["not_found"] },
          "message": { "type": "string", "examples": ["resource not found"] },
          "hint": {
            "type": "string",
            "examples": [
              "Check the path and consult https://www.fortyone.app/openapi.json."
            ]
          }
        }
      },
      "JSONRPCRequest": {
        "type": "object",
        "required": ["jsonrpc", "method"],
        "properties": {
          "jsonrpc": { "const": "2.0" },
          "id": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "method": { "type": "string" },
          "params": { "type": "object" }
        }
      }
    }
  }
}